upgrade project to VS2022
Hello Mixere! Welcome to the year 2021, where we no longer use Microsoft
Visual C++ 6. I know, everything is different and it feels weird, we'll
get through it.

This changeset migrates the project (via Visual Studio
Express 2008) to Visual Studio 2022. It also makes a few minor changes
to the source to build with the more strict compiler.

We also don't use CRLF line endings, multiple-document interface, or MFC
for that matter, but one change at a time.